python/requirements.txt
azure-cognitiveservices-speech>=1.40.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
